requirements.txt
evdev==1.2.0
PyYAML==5.1.1
